refactor: Reworked project structure
