--- .codespellrc
@@ -1,7 +1,0 @@
-[codespell]
-builtin = clear,rare,en-GB_to_en-US,names,informal,code
-check-filenames =
-check-hidden =
-skip = */.git,*/build,*/prefix,*/.codespellrc
-quiet-level = 2
-ignore-regex = ^#include <(?:stdio)\.h>$
