+++ .gitignore
@@ -1,15 +1,7 @@
-**/.DS_Store
-.idea/
-.vs/
-.vscode/
-build/
-cmake-build-*/
-prefix/
-.clangd
-CMakeLists.txt.user
-CMakeUserPresets.json
-compile_commands.json
-env.bat
-env.ps1
-.envrc
+.cache
 .direnv
+.envrc
+output
+**/*.o
+**/*.d
+compile_commands.json
