+++ .vscode/extensions.json
@@ -1,0 +1,7 @@
+{
+    "recommendations": [
+        "pkief.material-icon-theme",
+        "llvm-vs-code-extensions.vscode-clangd",
+        "kylinideteam.cppdebug"
+    ]
+}
