--- README.md
@@ -1,7 +1,0 @@
-# iobuf
-
-This is the iobuf project.
-
-# Building and installing
-
-See the [BUILDING](BUILDING.md) document.
