--- cmake/install-config.cmake
@@ -1,1 +1,0 @@
-include("${CMAKE_CURRENT_LIST_DIR}/iobufTargets.cmake")
