--- docs/conf.py.in
@@ -1,5 +1,0 @@
-DOXYFILE = 'Doxyfile'
-
-LINKS_NAVBAR1 = [
-    (None, 'pages', [(None, 'about')]),
-]
