+++ flake.nix
@@ -1,20 +1,31 @@
 {
-  description = "Best game";
+  description = "Cringe dev environment (it's C dude)";
 
-  inputs = {};
+  inputs = { };
 
-  outputs = {self, nixpkgs} @ inputs: let
-    system ="x86_64-linux";
-    pkgs = nixpkgs.legacyPackages.${system};
-  in {
-    devShells.${system}.default = with pkgs; mkShell rec {
-      packages = [
-        clang-tools
-        lldb
-        cmake
-        cmake-language-server
-        gdb
-      ];
+  outputs = { self, nixpkgs } @ inputs:
+    let
+      system = "x86_64-linux";
+      pkgs = nixpkgs.legacyPackages.${system};
+    in
+    {
+      devShells.${system}.default = with pkgs; mkShell rec {
+        packages = [
+          gcc
+          gdb
+          lldb
+          bear
+          clang
+          gnumake
+          clang-tools
+          nixpkgs-fmt
+          pkg-config
+        ];
+        buildInputs = [
+          gcc
+          clang
+        ];
+        LD_LIBRARY_PATH = lib.makeLibraryPath buildInputs;
+      };
     };
-  };
 }
