+++ src/main.c
@@ -1,7 +1,7 @@
+#include <stdio.h>
+
 int main(int argc, char const* argv[])
 {
-  (void)argc;
-  (void)argv;
-
+  printf("Hello World!\n");
   return 0;
 }
