+++ test/test-format.c
