+++ test/test-rand.c
